requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-cockroachdb==3.2.2
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
zipp==3.6.0
